Implement initial data management and plotting
Implemented an initial tree-based data reading/processing structure,
with distinct "reader" and "region" classes. This feels overcomplex, so
I might remove the tree structure and reader/region distinction next.

Scripted a line plot and choropleth plot in Jupyter Notebook, in order
to sketch out what needs to happen in a graphing object. The choropleth
plot is not fully functional yet, but does draw a map of the US.

The goal is the create a Jupyter Notebook-based web app with interactive
plots, via either Voila or Jupyter Books.