Finished data management & created state selection
I finally realized a simpler way to design the data management OOP:
instead of creating objects that represent regions, and have it be
responsible for all the data pertaining to that region, it's much more
straightforward to create objects that represent chunks of data; i.e.
instead of an object that represents, say, Pennsylvania as a whole, we
have an object that represents, say, county-level data of counties
within Pennsylvania.

Additionally, I noticed that the structure of these data objects is a
tree: on the top level, we have 3 different levels of data granularity
(all US, state-by-state, county-by-county). Each of these nodes has
many children, each of which represents a region (e.g. California,
Allegheny County, or, trivially, All US). Each of these nodes in turn
can be further split by date (October 1st, 2020 or September 27th,
2020).

Therefore, to exploit this underlying structure, I ordered the classes
inside each other like Russian nesting dolls: `Date` is an inner class
of `Region`, which in turn is an inner class of `Level`.

There are a few things that all of my classes shared. In order to keep
my code exceptionally DRY, I created a `Data` abstract base class:
- Each instance one-to-one represents a chunk of data. Therefore, each
   class should simply just need to define how to create that data from
   initialization parameters, and the `Data` parent class handles the
   rest.
- The calculation of data should be lazy; the data property should not
   be generated until it's needed.
- The instances should be parameter-dependent singletons; i.e. at most
   one instance per combination of class and initialization parameters.
   I tried out building a metaclass to make this happen but I realized
   that it's possible to simply use (abuse) the `functools.lru_cache`
   decorator to memoize the Python `__new__()` magic function.
- I wanted the notation to be very clean, with a lot of method chaining
   but without extraneous keywords like ".data()". Therefore, it is
   possible to get the data that each instance represents by calling it
   like a function, via `__call__()` overrides. Also, there are a few
   convenience functions that simplify initializing nested instances to
   allow for chaining. The result is that one-line expressions such as
   `county.region("42003").date("2020-09-27")()` are possible.

Meanwhile in `app.ipynb`, I've figured out how to create a map of the US
states where clicking on a state will display a corresponding map of the
counties within that state. The next step is to wrap up these plots into
creator objects.