Created geoJSON and choropleth manager objects
The `GeoJSONManager` class is a child class of Data, but specialized for
managing the GeoJSON geographic data files that power the choropleth
plots. Since this is JSON, it has to make some very strong assumptions
about the structure of the data. However, Plotly is very strict with its
GeoJSON specifications anyway, so I don't think that will be a problem.

The `Choropleth` class is a child class of an abstract `Figure` manager
class, and creates, manages, and update interactive plotly choropleth
traces and figures. A basic implementation is working in app.ipynb.
Cleaning that notebook up is the next step.