Completed Jupyter Notebook dashboard
The full interactivity of the dashboard has been implemented, along with
some javascript-Jupyter magic to hide all the code and run everything
automatically.

The region of line and choropleth plots are selected by clicking on
the choropleth regions, and the date of the choropleth figures are
selected by clicking on the line plots. Additionally, a toggle switch
swaps the choropleth plots between showing case and death data. The
choropleth plots tend to be kind of glitchy, especially the
nation-wide, state-by-state one. I don't think that it's on my end, but
just Jupyter Notebooks being badly supported by Plotly now that they
have launched their proprietary Dash dashboarding tool. I provided as
much support I can in the text of the Notebook as I can, but the
nation-wide choropleth will just disappear randomly or refuse to update.
Barring that, this worked suprisingly well.

The next step would be to finally write a `main.py`, finalize cleaning
up the code, and document the many different OOP techniques involved
here.